+++ gdsfactory/__init__.py
@@ -27,7 +27,6 @@ from gdsfactory.component import Component, ComponentReference
 from gdsfactory.config import CONF, call_if_func, PATH, logger
 from gdsfactory.port import Port
 from gdsfactory.cell import cell
-from gdsfactory.cell import declarative_cell
 from gdsfactory.cell import cell_without_validator
 from gdsfactory.cell import clear_cache
 from gdsfactory.show import show
@@ -101,7 +100,6 @@ __all__ = (
     "c",
     "call_if_func",
     "cell",
-    "declarative_cell",
     "cell_without_validator",
     "clear_cache",
     "components",
